test(core): split pipeline-m1 and pipeline-geometry characterization suites
Replace the 718-line pipeline-m1 and 673-line pipeline-geometry monoliths
with focused suites under tests/pipeline-m1/ and tests/pipeline-geometry/,
grouped by geom/scales/theme/failure policy and by helper concern, with
shared fixtures.
