+++ packages/core/tests/pipeline-geometry/fixtures.ts
@@ -1,0 +1,4 @@
+/**
+ * Shared fixtures for pipeline geometry characterization.
+ */
+export const size = { width: 640, height: 400 };
