+++ packages/core/tests/pipeline-m1/fixtures.ts
@@ -1,0 +1,11 @@
+/**
+ * Shared fixtures for M1 pipeline surface characterization.
+ */
+export const size = { width: 640, height: 400 };
+
+export const salesRows = [
+  { city: "Berlin", kind: "web", sales: 30 },
+  { city: "Berlin", kind: "store", sales: 20 },
+  { city: "Oslo", kind: "web", sales: 10 },
+  { city: "Oslo", kind: "store", sales: 25 },
+];
